Introduced plugin framework. Reworked what config files do.
Most of the functionality previously added in config, and defined in the qax config json files has been moved to the plugin definition. Reason being is that it will be the check tools that define what inputs (files and params) they support, and this information should be defined in the check tool and not in the qax configuration file.

confiog files now really only contain a list of check tools and the classpaths to load the check tool plugins from