includes status for check execution
Addresses usuability issue whereby user would click a button, and it would then not be clearly indicated what was happening. Eg; clicking the stop button would stop the checks, but there may be a delay in doing so. Previously the check would have continued running, and stopped eventually. Now at least a 'stopping' status is displayed, this then changes to stopped when the check has actually stopped.